Refactor and remove redundant components
